Rearchitected robot and downloader
Rewrote Downloader to be more independent of Robot and RequestPDFScreen,
such that they are more loosely coupled by an event system.  One of the
motivations for the change is to include information of defective PDFs
in the final summary.  The Summarizer is now a separate class that keeps
track of what happens by listening to events.